+++ main.js
@@ -3,10 +3,6 @@ require('isomorphic-fetch');
 const fs = require('fs');
 const path = require('path');
 const express = require('express');
-const http = require('http');
-const https = require('https');
-const denodeify = require('denodeify');
-
 const flags = require('next-feature-flags-client');
 const backendAuthentication = require('./src/middleware/backend-authentication');
 
@@ -14,184 +10,27 @@ const backendAuthentication = require('./src/middleware/backend-authentication')
 const metrics = require('next-metrics');
 const nLogger = require('@financial-times/n-logger').default;
 const serviceMetrics = require('./src/lib/service-metrics');
-const raven = require('@financial-times/n-raven');
-const healthChecks = require('./src/lib/health-checks');
+
 
 // utils
-const normalizeName = require('./src/lib/normalize-name');
+const healthChecks = require('./src/lib/health-checks');
+const instrumentListen = require('./src/lib/instrument-listen');
+const guessAppDetails = require('./src/lib/guess-app-details');
+
 const robots = require('./src/middleware/robots');
 const vary = require('./src/middleware/vary');
 const cache = require('./src/middleware/cache');
 
 // Health check failure simulation
 const checkFailing = require('./src/lib/check-failing');
+
 const teapot = fs.readFileSync(path.join(__dirname, 'src/assets/teapot.ascii'), 'utf8');
 
-const guessAppDetails = options => {
-	let packageJson = {};
-	let name = options.name;
-	let description = '';
-	let directory = options.directory || process.cwd();
 
-	if (!name) {
-		try {
-			packageJson = require(directory + '/package.json');
-			name = packageJson.name;
-			description = packageJson.description || '';
-		} catch(e) {
-			// Safely ignorable error
-		}
-	}
+const getAppContainer = options => {
 
-	if (!name) throw new Error('Please specify an application name');
-
-	name = name && normalizeName(name);
-
-	return {name, description, directory};
-}
-
-
-function chain () {
 	checkFailing.init();
 
-	const app = express();
-	const initPromises = [];
-
-	const actualAppListen = function () {
-		let serverPromise;
-		if (process.argv.indexOf('--https') > -1) {
-			const readFile = denodeify(fs.readFile);
-			serverPromise = Promise.all([
-				readFile(path.resolve(__dirname, 'key.pem')),
-				readFile(path.resolve(__dirname, 'cert.pem'))
-			])
-				.then(results => https.createServer({ key: results[0], cert: results[1] }, this));
-		} else {
-			serverPromise = Promise.resolve(http.createServer(this));
-		}
-
-		return serverPromise.then(server => server.listen.apply(server, arguments));
-	};
-
-	app.listen = function () {
-		const args = [].slice.apply(arguments);
-		app.use(raven.middleware);
-		const port = args[0];
-		const cb = args[1];
-		args[1] = function () {
-			// HACK: Use warn so that it gets into Splunk logs
-			nLogger.warn({ event: 'EXPRESS_START', app: app.nextMeta.name, port: port, nodeVersion: process.version });
-			return cb && cb.apply(this, arguments);
-		};
-
-		return Promise.all(initPromises)
-			.then(() => {
-				metrics.count('express.start');
-				return actualAppListen.apply(app, args);
-			})
-			// Crash app if initPromises fail
-			.catch(err => setTimeout(() => {
-				throw err;
-			}));
-	};
-
-	return Object.create(chainConstructor, {
-		app: {value: app},
-		addInitPromise: {value: initPromises.push.bind(initPromises)}
-	});
-}
-
-const chainConstructor = {
-	config: function (options) {
-		//Remove x-powered-by header
-		this.app.set('x-powered-by', false);
-
-		this.app.get('/robots.txt', robots);
-
-		this.app.get('/__brew-coffee', (req, res) => {
-			res.status(418);
-			res.send(teapot);
-			res.end();
-		});
-
-		// utility middleware
-		this.app.use(cache);
-		this.app.use(vary);
-		this.app.nextMeta = guessAppDetails(options)
-		return this;
-	},
-
-	apply: function (func) {
-		return func.call(this);
-	},
-
-	healthChecks: function (systemCode, checks) {
-		healthChecks(this.app, {systemCode, healthChecks: checks}, this.app.nextMeta.description)
-		return this;
-	},
-
-	about: function () {
-		this.app.get('/__about', (req, res) => {
-			res.set({ 'Cache-Control': 'no-cache' });
-			res.sendFile(this.app.nextMeta.directory + '/public/__about.json');
-		});
-		return this;
-	},
-
-	metrics: function (downstreamMetrics) {
-		// metrics should be one of the first things as needs to be applied before any other middleware executes
-		metrics.init({ app: this.app.nextMeta.name, flushEvery: 40000 });
-		this.app.use((req, res, next) => {
-			metrics.instrument(req, { as: 'express.http.req' });
-			metrics.instrument(res, { as: 'express.http.res' });
-			next();
-		});
-
-		if (downstreamMetrics) {
-			serviceMetrics.init();
-		}
-		return this;
-	},
-
-	timestamp: function () {
-		this.app.use((req, res, next) => {
-			res.set('FT-Backend-Timestamp', new Date().toISOString());
-			next();
-		});
-
-		return this;
-	},
-
-	backendAuth: function (withAuth) {
-		// Only allow authorized upstream applications access
-		if (withAuth) {
-			this.app.use(backendAuthentication(this.app.nextMeta.name));
-		} else {
-			nLogger.warn({ event: 'BACKEND_AUTHENTICATION_DISABLED', message: 'Backend authentication is disabled, this app is exposed directly to the internet' });
-		}
-		return this;
-	},
-
-	flags: function (withFlags) {
-		// feature flags
-		if (withFlags) {
-			this.addInitPromise(flags.init());
-			this.app.use(flags.middleware);
-		}
-		return this;
-	},
-
-	withEverything: function () {
-		return this
-			.about()
-			.metrics(true)
-			.timestamp()
-			.backendAuth(true)
-			.flags(true)
-	}
-}
-
-module.exports = options => {
 	options = options || {};
 
 	const defaults = {
@@ -207,25 +46,70 @@ module.exports = options => {
 		}
 	});
 
-	// When adding to the chain below please consider whether n-ui
-	// (Which wraps n-express and calls these methods in a different order)
-	// needs to be updated too
-	const app = chain()
-		.config(options)
-		.healthChecks(options.systemCode, options.healthChecks)
-		.about()
-		.metrics(options.withServiceMetrics)
-		.timestamp()
-		.backendAuth(options.withBackendAuthentication)
-		.flags(options.withFlags)
-		.app;
+	const meta = guessAppDetails(options);
+	const initPromises = [];
+	const app = instrumentListen(express(), meta, initPromises);
+	const addInitPromise = initPromises.push.bind(initPromises)
 
-	return app;
+	//Remove x-powered-by header
+	app.set('x-powered-by', false);
+
+	app.get('/robots.txt', robots);
+
+	app.get('/__brew-coffee', (req, res) => {
+		res.status(418);
+		res.send(teapot);
+		res.end();
+	});
+
+	// utility middleware
+	app.use(cache);
+	app.use(vary);
+
+	healthChecks(app, options, meta.description)
+
+	app.use((req, res, next) => {
+		res.set('FT-Backend-Timestamp', new Date().toISOString());
+		next();
+	});
+
+	// metrics should be one of the first things as needs to be applied before any other middleware executes
+	metrics.init({ app: meta.name, flushEvery: 40000 });
+	app.use((req, res, next) => {
+		metrics.instrument(req, { as: 'express.http.req' });
+		metrics.instrument(res, { as: 'express.http.res' });
+		next();
+	});
+
+	if (options.withServiceMetrics) {
+		serviceMetrics.init();
+	}
+
+	app.get('/__about', (req, res) => {
+		res.set({ 'Cache-Control': 'no-cache' });
+		res.sendFile(meta.directory + '/public/__about.json');
+	});
+
+	if (options.withBackendAuthentication) {
+		app.use(backendAuthentication(meta.name));
+	} else {
+		nLogger.warn({ event: 'BACKEND_AUTHENTICATION_DISABLED', message: 'Backend authentication is disabled, this app is exposed directly to the internet' });
+	}
+
+	// feature flags
+	if (options.withFlags) {
+		addInitPromise(flags.init());
+		app.use(flags.middleware);
+	}
+
+	return { app,	meta,	addInitPromise };
 };
+
+module.exports = options => getAppContainer(options).app
 
 // expose internals the app may want access to
 module.exports.Router = express.Router;
 module.exports.static = express.static;
 module.exports.metrics = metrics;
 module.exports.flags = flags;
-module.exports.chain = chain;
+module.exports.getAppContainer = getAppContainer;
